+++ src/main/java/ch/admin/bj/swiyu/issuer/management/api/exception/RestExceptionHandler.java
@@ -12,7 +12,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
@@ -24,6 +23,7 @@ import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExcep
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static org.springframework.http.HttpStatus.*;
@@ -78,13 +78,14 @@ public class RestExceptionHandler extends ResponseEntityExceptionHandler {
                                                                   @NonNull HttpHeaders headers,
                                                                   @NonNull HttpStatusCode status,
                                                                   @NonNull WebRequest request) {
-        List<String> errors = Stream.concat(
-                ex.getBindingResult().getFieldErrors()
-                        .stream().map(error -> String.format("%s: %s", error.getField(), error.getDefaultMessage())),
-                ex.getBindingResult().getGlobalErrors().stream().map(error -> String.format("%s: %s", error.getObjectName(), error.getDefaultMessage()))
-        ).toList();
+        String errors = Stream.concat(
+                        ex.getBindingResult().getFieldErrors()
+                                .stream().map(error -> String.format("%s: %s", error.getField(), error.getDefaultMessage())),
+                        ex.getBindingResult().getGlobalErrors().stream().map(error -> String.format("%s: %s", error.getObjectName(), error.getDefaultMessage()))
+                ).sorted()
+                .collect(Collectors.joining(", "));
 
-        return new ResponseEntity<>(getErrorsMap(errors), new HttpHeaders(), HttpStatus.BAD_REQUEST);
+        return new ResponseEntity<>(errors, BAD_REQUEST);
     }
 
     private Map<String, List<String>> getErrorsMap(List<String> errors) {