+++ src/main/java/ch/admin/bj/swiyu/issuer/management/api/statuslist/StatusListConfigDto.java
@@ -6,7 +6,6 @@
 
 package ch.admin.bj.swiyu.issuer.management.api.statuslist;
 
-import ch.admin.bj.swiyu.issuer.management.api.validators.ValidBits;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
@@ -18,7 +17,7 @@ public class StatusListConfigDto {
     private String purpose;
 
     @NotNull
-    @ValidBits
+    @ValidStatusListBits
     @Schema(description = "The number of bits used per Referenced Token. Possible values are 1, 2, 4, 8.", example = "2")
     private Integer bits;
 }