+++ src/test/java/ch/admin/bj/swiyu/issuer/management/statuslist/StatusListMaxLengthValidatorTest.java
@@ -9,7 +9,7 @@ package ch.admin.bj.swiyu.issuer.management.statuslist;
 import ch.admin.bj.swiyu.issuer.management.api.statuslist.StatusListConfigDto;
 import ch.admin.bj.swiyu.issuer.management.api.statuslist.StatusListCreateDto;
 import ch.admin.bj.swiyu.issuer.management.api.statuslist.StatusListTypeDto;
-import ch.admin.bj.swiyu.issuer.management.api.validators.ValidStatusListMaxLengthValidator;
+import ch.admin.bj.swiyu.issuer.management.api.statuslist.ValidStatusListMaxLengthValidator;
 import ch.admin.bj.swiyu.issuer.management.common.config.StatusListProperties;
 import jakarta.validation.ConstraintValidatorContext;
 import org.junit.jupiter.api.BeforeEach;