+++ Changelog.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.3.7
+## 1.4.0
 
 ### Changed
 
-- Update and improved the error feedback in the status list create dto. Also added a list of errors to the response.
+- Update and improved the error feedback in the status list create dto. The error messages are now more detailed and
+  provide more information about the error.
 
 ## 1.3.6